requirements.txt
fastapi
uvicorn[standard]
telethon
opentele
python-multipart
requests
pydantic
python-dotenv
aiofiles
jinja2
cryptography